requirements.txt
anyio==3.6.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
dateparser==1.1.8
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
line-profiler==4.0.3
lxml==4.9.2
pydantic==1.10.7
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
regex==2023.3.23
requests==2.28.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.9
starlette==0.26.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.13
uvicorn==0.21.1
yapf==0.32.0
